Change field types

+++ internal/resource/dto.go
@@ -1,8 +1,8 @@
 package resource
 
 type resourceDTO struct {
-	ID           int64
-	Version      int64
+	ID           uint
+	Version      int
 	TextField    *string
 	NumberField  *int
 	BooleanField *bool
